Added darwin/Makefile to generate an universal binary for Mac OS X by only running make. This currently generates an universal binary with JIT support for Intel Macs, although it currently crashes. Changed the build system to allow two builds for different targets, which is needed by the new darwin/Makefile. Further changes have been done to mainly solve compiler warnings.

+++ src/ndebug.cpp
@@ -292,7 +292,7 @@ void ndebug::m68k_print(FILE * f)
 		   (int) (regs.dfc % 2));
 
     for (unsigned int i = 0; i < 8; i++) {
-	fprintf(f, "FP%d: %Lg ", i, fpu.registers[i]);
+	fprintf(f, "FP%d: %Lg ", i, (long double)fpu.registers[i]);
 	if (i == 3) fprintf(f, "N=%d Z=%d\n",
 	    (fpu.fpsr.condition_codes & 0x8000000) != 0,
 	    (fpu.fpsr.condition_codes & 0x4000000) != 0);
